Finishing implementation of Java version

+++ build.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
+
+if [ -f interpreter.js ]; then
+    rm interpreter.js
+fi
+
 FILES=`find src -name "*.js"`
 for FILE in $FILES; do cat $FILE >> interpreter.js; done
 
 yui -o interpreter.min.js interpreter.js
-rm interpreter.js
